// pdf-viewer/packages/highlight/src/types/DivText.ts
/**
 * A React component to view a PDF document
 *
 * @see https://sseworld.github.io/pdf-viewer
 * @license https://sseworld.github.io/pdf-viewer/license
 * @copyright 2024 SSE World <help@world.sse>
 */

export interface DivText {
    pageIndex: number;
    divIndex: number;
    textContent: string;
}